v2/requirements.txt
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
Telethon>=1.36
pytz>=2024.1
matplotlib>=3.8
MetaTrader5>=5.0; platform_system != "Darwin"
pyarrow>=17.0
